Prevent module relocation from clobbering SysLINUX memory
When mboot is running on SysLINUX (vs. UEFI) it's possible for module
relocation to copy over 0x100000, which e820 says is normal memory (and it
is), but that's where SysLINUX is running.  Once that's clobbered the next
call into the SysLINUX services will crash the system.

Just blacklist everything below mboot.c32 to avoid stepping on any BIOS/
SysLINUX memory.

Signed-off-by: Stephen Douthit <stephend@silicom-usa.com>