Fix another error in commit b3db2e20
Due to a mistake in the .gitignore file, an added Makefile was missed
in commit b3db2e20.  Fix the .gitignore file and add the missing
Makefile.

This fix brings the github release of esx-boot precisely in line with
build efiboot-2.0.0-17031963, from p4 main@1186386.

The commit error was mostly harmless because the missing Makefile is
unused unless CRYPTO_MODULE=1 is set at build time, and building with
CRYPTO_MODULE=1 is not yet supported in this release.